pyproject.toml
[project]
name = "immich-upload-daemon"
version = "0.1.7"
description = "Immich Upload Daemon is a tool designed to monitor your media directories for new image and video files and automatically upload them to an Immich server."
readme = "README.md"
authors = [
    { name = "Luis Garcia", email = "git@luigi311.com" }
]

requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.0",
    "imohash>=1.1.0",
    "loguru>=0.7.3",
    "pyxdg>=0.28",
    "sdbus-networkmanager>=2.0.0",
    "uvloop>=0.21.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "nuitka>=2.7.12",
    "ruff>=0.11.2",
    "types-aiofiles>=24.1.0.20241221",
    "types-pyxdg>=0.28.0.20240106",
]

[project.scripts]
immich_upload_daemon = "immich_upload_daemon.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

